backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==6.0.1
python-socketio==5.9.0
python-engineio==4.7.1
simple-websocket==1.0.0
PyJWT==2.8.0
python-dotenv==1.0.0